pyproject.toml
[tool.poetry]
name = "bigmeow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
structlog = "^24.2.0"
python-telegram-bot = "^21.2"
python-cowsay = "^1.2.0"
discord-py = "^2.3.2"
fastapi = {extras = ["standard"], version = "^0.114.2"}
aiohttp = {extras = ["speedups"], version = "^3.10.5"}

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
prospector = "^1.10.3"
pytest = "^8.2.1"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
